Split Transformer, PartialTransformer and Patcher into shared code and *Platform files
