+++ .scalafmt.conf
@@ -2,6 +2,11 @@ version = 3.5.9
 project.git = true
 maxColumn = 120
 runner.dialect = Scala213Source3
+fileOverride {
+  "glob:**/chimney/src/main/scala-3/**" {
+    runner.dialect = scala3
+  }
+}
 
 align.preset = some
