+++ chimney/src/main/scala-2/io/scalaland/chimney/PartialTransformerCompanionPlatform.scala
@@ -1,0 +1,20 @@
+package io.scalaland.chimney
+
+import io.scalaland.chimney.internal.macros.dsl.TransformerBlackboxMacros
+
+import scala.language.experimental.macros
+
+private[chimney] trait PartialTransformerCompanionPlatform { this: PartialTransformer.type =>
+
+  /** Provides [[io.scalaland.chimney.PartialTransformer]] derived with the default settings.
+   *
+   * When transformation can't be derived, it results with compilation error.
+   *
+   * @tparam From type of input value
+   * @tparam To   type of output value
+   * @return [[io.scalaland.chimney.PartialTransformer]] type class definition
+   * @since 0.7.0
+   */
+  implicit def derive[From, To]: PartialTransformer[From, To] =
+    macro TransformerBlackboxMacros.derivePartialTransformerImpl[From, To]
+}
