+++ chimney/src/main/scala/io/scalaland/chimney/PartialTransformer.scala
@@ -1,10 +1,7 @@
 package io.scalaland.chimney
 
 import io.scalaland.chimney.dsl.{PartialTransformerDefinition, TransformerDefinitionCommons}
-import io.scalaland.chimney.internal.macros.dsl.TransformerBlackboxMacros
 import io.scalaland.chimney.internal.{TransformerCfg, TransformerFlags}
-
-import scala.language.experimental.macros
 
 /** Type class expressing partial transformation between
   * source type `From` and target type `To`, with the ability
@@ -48,7 +45,7 @@ trait PartialTransformer[From, To] {
     transform(src, failFast = true)
 }
 
-object PartialTransformer {
+object PartialTransformer extends PartialTransformerCompanionPlatform {
 
   /** Construct ad-hoc instance of partial transformer from transforming function returning partial result.
     *
@@ -97,19 +94,6 @@ object PartialTransformer {
     */
   def liftTotal[From, To](t: Transformer[From, To]): PartialTransformer[From, To] =
     fromFunction[From, To](t.transform)
-
-  /** Provides [[io.scalaland.chimney.PartialTransformer]] derived with the default settings.
-    *
-    * When transformation can't be derived, it results with compilation error.
-    *
-    * @tparam From type of input value
-    * @tparam To type of output value
-    * @return [[io.scalaland.chimney.PartialTransformer]] type class definition
-    *
-    * @since 0.7.0
-    */
-  implicit def derive[From, To]: PartialTransformer[From, To] =
-    macro TransformerBlackboxMacros.derivePartialTransformerImpl[From, To]
 
   /** Creates an empty [[io.scalaland.chimney.dsl.PartialTransformerDefinition]] that
     * you can customize to derive [[io.scalaland.chimney.PartialTransformer]].
