+++ chimney/src/main/scala/io/scalaland/chimney/Transformer.scala
@@ -1,10 +1,7 @@
 package io.scalaland.chimney
 
-import io.scalaland.chimney.internal.{TransformerCfg, TransformerFlags}
 import io.scalaland.chimney.dsl.{PartialTransformerDefinition, TransformerDefinition, TransformerDefinitionCommons}
-import io.scalaland.chimney.internal.macros.dsl.TransformerBlackboxMacros
-
-import scala.language.experimental.macros
+import io.scalaland.chimney.internal.{TransformerCfg, TransformerFlags}
 
 /** Type class expressing total transformation between
   * source type `From` and target type `To`.
@@ -26,20 +23,7 @@ trait Transformer[From, To] {
   def transform(src: From): To
 }
 
-object Transformer {
-
-  /** Provides [[io.scalaland.chimney.Transformer]] derived with the default settings.
-    *
-    * When transformation can't be derived, it results with compilation error.
-    *
-    * @tparam From type of input value
-    * @tparam To type of output value
-    * @return [[io.scalaland.chimney.Transformer]] type class instance
-    *
-    * @since 0.2.0
-    */
-  implicit def derive[From, To]: Transformer[From, To] =
-    macro TransformerBlackboxMacros.deriveTransformerImpl[From, To]
+object Transformer extends TransformerCompanionPlatform {
 
   /** Creates an empty [[io.scalaland.chimney.dsl.TransformerDefinition]] that
     * you can customize to derive [[io.scalaland.chimney.Transformer]].
